Refactor high score data types to use Double for click speed; Changed friends only scoreboard query to display user even if they have 0 friends

+++ src/main/java/ch/uzh/ifi/hase/soprafs26/rest/dto/HighScoresDTO.java
@@ -6,7 +6,7 @@ public class HighScoresDTO { //data sent when updating high scores
     private int[] typingScores;
     private double[] timeIntervalScores;
     private int[] aimTestScores;
-    private int[] clickSpeedScores;
+    private double[] clickSpeedScores;
 
     public int[] getReactionScores() {
         return reactionScores;
@@ -40,11 +40,11 @@ public class HighScoresDTO { //data sent when updating high scores
         this.aimTestScores = aimTestScores;
     }
 
-    public int[] getClickSpeedScores() {
+    public double[] getClickSpeedScores() {
         return clickSpeedScores;
     }
 
-    public void setClickSpeedScores(int[] clickSpeedScores) {
+    public void setClickSpeedScores(double[] clickSpeedScores) {
         this.clickSpeedScores = clickSpeedScores;
     }
