+++ src/main/java/ch/uzh/ifi/hase/soprafs26/service/UserService.java
@@ -189,7 +189,7 @@ public class UserService {
 	}
 
 	public HighScoresResponseDTO updateHighScores(Long id, int[] reactionScores, int[] typingScores,
-			double[] timeIntervalScores, int [] aimTestScores, int[] clickSpeedScores) {
+			double[] timeIntervalScores, int [] aimTestScores, double[] clickSpeedScores) {
 		User user = userRepository.findById(id)
 			.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, 
 			"User with id [" + id + "] not found"));
@@ -261,8 +261,8 @@ public class UserService {
 
 		// Update click speed high score (higher is better, so check maximum)
 		if (clickSpeedScores != null && clickSpeedScores.length > 0) {
-			int maxClickSpeedScore = 0;
-			for (int score : clickSpeedScores) {
+			double maxClickSpeedScore = 0;
+			for (double score : clickSpeedScores) {
 				if (score > maxClickSpeedScore) {
 					maxClickSpeedScore = score;
 				}
